Made frontend separate module like backend
